chore: replace nordpool cron jobs with signle function
