+++ ui/api-utils/utils.js
@@ -1,0 +1,51 @@
+export const STATIC_SPOTPRICE_PARAMS = {
+  market: "DayAhead",
+  deliveryAreas: ["SE1", "SE2", "SE3", "SE4"],
+  currency: "SEK",
+};
+
+export const PRICE_AREAS = ["SE1", "SE2", "SE3", "SE4"];
+
+export const COLOR_LIST = [
+  "#99621E",
+  "#D38B5D",
+  "#F3FFB6",
+  "#739E82",
+  "#2C5530",
+  "#A67C52",
+  "#E4B07A",
+  "#FAEECF",
+  "#8AA899",
+  "#3D6653",
+  "#B0885C",
+  "#D9A87C",
+  "#F2E8C6",
+  "#7D9A87",
+  "#466A5D",
+  "#C4A484",
+  "#E0C3A2",
+  "#FFF7E4",
+  "#99B3A6",
+  "#5B7F6B",
+];
+
+export const LABEL_DICTIONARY = {
+  HydroWaterReservoir: "Vattenkraftmagasin",
+  Other: "Övrigt",
+  WindOnshore: "Vindkraft land",
+  Nuclear: "Kärnkraft",
+  FossilGas: "Fossilgas",
+  Biomass: "Biomassa",
+  FossilHardCoal: "Stenkol",
+  FossilOil: "Fossilolja",
+  HydroRunOfRiverAndPoundage: "Vattenkraft",
+  OtherRenewable: "Annan förnybar energi",
+  Solar: "Solkraft",
+  Waste: "Avfall",
+  WindOffshore: "Vindkraft havsbaserad",
+  FossilPeat: "Fossilt torv",
+  HydroPumpedStorage: "Vattenkraft (pumpkraft)",
+  FossilBrownCoal: "Brunkol",
+  Geothermal: "Geotermisk energi",
+  FossilCoalDerivedGas: "Kolbaserad gas",
+};
