+++ ui/vercel.json
@@ -1,15 +1,7 @@
 {
   "crons": [
     {
-      "path": "/api/cron-functions/updateSpotPrices",
-      "schedule": "5 2 * * *"
-    },
-    {
-      "path": "/api/cron-functions/updateEnergyMix",
-      "schedule": "5 2 * * *"
-    },
-    {
-      "path": "/api/cron-functions/updateExchange",
+      "path": "/api/cron-functions/updateNordpoolData",
       "schedule": "5 2 * * *"
     },
     {
