--- vercel.json
@@ -1,20 +1,0 @@
-{
-  "crons": [
-    {
-      "path": "/api/cron-functions/updateSpotPrices",
-      "schedule": "5 2 * * *"
-    },
-    {
-      "path": "/api/cron-functions/updateEnergyMix",
-      "schedule": "5 2 * * *"
-    },
-    {
-      "path": "/api/cron-functions/updateExchange",
-      "schedule": "5 2 * * *"
-    },
-    {
-      "path": "/api/cron-functions/updateAiAnalysis",
-      "schedule": "15 2 * * *"
-    }
-  ]
-}
